code move DockerController to separate module
